+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added resolver.py which resolves a parsed reference to obtain the full context of the transformation.
 - Added fetcher.py, ported from transform.py, to allow the use of existing fethcing/compositing for use in transformations.
 - Added CLI commands the help track the progress of porting references from definitions.
+- Added `engine` module.
+- Added `grid` module.
 
 ### Changed
 - Updated CLI options to use new dist2coast km decay and blend options
@@ -38,8 +40,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved generated-grid cache identity to account for transformation references, epochs, region, resolution, decay settings, and other generation options.
 - Moved vertical unit metadata out of the legacy datum registry and into the typed reference model.
 - Refactored to adhere to ruff PTH rules.
-- Moved htdp and vdatum to /engines
-- Moved fetchez modules and hooks to /fetchez/...
+- Moved htdp and vdatum to `/engines`
+- Moved fetchez modules and hooks to `/fetchez/...`
+- Moved `GridWriter` into new `/grid/io` module
+- Moved `generation.py` to `/grid/shift` module
+- Moved `grid_engine.py` to `/grid/engine` module
 
 ### Fixed
 - Refactor dist2coast usage to fix a bug that would burn the dist2coast edges onto the raster result. This was due to performing an edt distance transform from the low-res dist2coast raster, treating it as a mask instead of a distance field. Update allows setting the distance by km instead of number of pixels and smooths the 'zero' field to get proper transitions.